+++ src/DfWriter.cpp
@@ -2,6 +2,7 @@
 #include "haven_types.h"
 #include "tagged_na.h"
 
+#include <unordered_map>
 #include "cpp11/doubles.hpp"
 #include "cpp11/strings.hpp"
 #include "cpp11/integers.hpp"
@@ -63,13 +64,16 @@ inline int userWidth(cpp11::sexp x) {
 class Writer {
   FileExt ext_;
   FileVendor vendor_;
+  int version_;
+  int strl_threshold_;
+  std::unordered_map<const char*, readstat_string_ref_t*> string_ref_;
 
   cpp11::list x_;
   readstat_writer_t* writer_;
   FILE* pOut_;
 
 public:
-  Writer(FileExt ext, cpp11::list x, cpp11::strings pathEnc): ext_(ext), vendor_(extVendor(ext)), x_(x) {
+  Writer(FileExt ext, cpp11::list x, cpp11::strings pathEnc): ext_(ext), vendor_(extVendor(ext)), version_(0), x_(x) {
     std::string path(Rf_translateChar(pathEnc[0]));
 
     pOut_ = fopen(path.c_str(), "wb");
@@ -92,6 +96,7 @@ public:
   }
 
   void setVersion(int version) {
+    version_ = version;
     readstat_writer_set_file_format_version(writer_, version);
   }
 
@@ -104,6 +109,10 @@ public:
       return;
 
     readstat_writer_set_file_label(writer_, string_utf8(label, 0));
+  }
+
+  void setStrLThreshold(int strl_threshold) {
+    strl_threshold_ = strl_threshold;
   }
 
   void write() {
@@ -369,8 +378,23 @@ public:
     }
 
 
-    readstat_variable_t* var =
-      readstat_add_variable(writer_, name, READSTAT_TYPE_STRING, user_width);
+    // Use strL for "long" strings in stata. strL has an 80 byte overhead so
+    // we use it when it's likely to be more efficient. The main downside of
+    // strL is that it can't be used as a join key but this seems unlikely for
+    // very long strings.
+    readstat_variable_t* var;
+    if (ext_ == HAVEN_DTA && version_ >= 117 && user_width > strl_threshold_) {
+      var = readstat_add_variable(writer_, name, READSTAT_TYPE_STRING_REF, user_width);
+      for (int i = 0; i < x.size(); ++i) {
+        const char* val = string_utf8(x, i);
+        if (!string_ref_.count(val)) {
+          string_ref_[val] = readstat_add_string_ref(writer_, val);
+        }
+      }
+    } else {
+      var = readstat_add_variable(writer_, name, READSTAT_TYPE_STRING, user_width);
+    }
+
     readstat_variable_set_format(var, format);
     readstat_variable_set_label(var, var_label(x));
     readstat_variable_set_label_set(var, labelSet);
@@ -426,6 +450,8 @@ public:
   readstat_error_t insertValue(readstat_variable_t* var, const char* val, bool is_missing) {
     if (is_missing) {
       return readstat_insert_missing_value(writer_, var);
+    } else if (var->type == READSTAT_TYPE_STRING_REF) {
+      return readstat_insert_string_ref(writer_, var, string_ref_[val]);
     } else {
       return readstat_insert_string_value(writer_, var, val);
     }
@@ -461,10 +487,11 @@ void write_sav_(cpp11::list data, cpp11::strings path, std::string compress) {
 }
 
 [[cpp11::register]]
-void write_dta_(cpp11::list data, cpp11::strings path, int version, cpp11::sexp label) {
+void write_dta_(cpp11::list data, cpp11::strings path, int version, cpp11::sexp label, int strl_threshold) {
   Writer writer(HAVEN_DTA, data, path);
   writer.setVersion(version);
   writer.setFileLabel(label);
+  writer.setStrLThreshold(strl_threshold);
   writer.write();
 }